BankChurn-Predictor/requirements.txt
# BankChurn Predictor - Production Dependencies
# Core Data Science Stack
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0

# Machine Learning Frameworks
scikit-learn>=1.0.0
xgboost>=1.5.0

# Hyperparameter Optimization
optuna>=3.0.0

# API Framework
fastapi>=0.78.0
uvicorn>=0.18.0
pydantic>=1.9.0

# Configuration Management
pyyaml>=6.0

# Model Persistence & Serialization
joblib>=1.1.0

# Data Validation & Type Hints
typing-extensions>=4.0.0

# Logging & Monitoring
python-multipart>=0.0.5

# Development & Testing (optional)
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Jupyter Environment
jupyter>=1.0.0
ipython>=7.0.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Progress Bars
tqdm>=4.62.0

# HTTP Requests (for API testing)
requests>=2.27.0

# MLOps / Monitoring (optional)
mlflow>=2.9
evidently>=0.4.8
